Moved the maxwell examples into the fem folder.
