+++ examples/fem/maxwell/multi_material/mesh.py
@@ -1,4 +1,3 @@
-import numpy as np
 import gmsh
 import sys
